rpc-server/requirements.txt
# pymongo>=4.0
# python-dotenv>=1.0

pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
datetime     
os-sys  